update: int64 fields updated to float64 because of db conflict

+++ internal/core/domain/customers.go
@@ -2,12 +2,12 @@ package domain
 
 // Customer is an entity handling commerce customers
 type Customer struct {
-	ID                string `db:"id"`
-	UserDNI           string `db:"user_dni"`
-	Name              string `db:"name"`
-	LastName          string `db:"last_name"`
-	Phone             string `db:"phone"`
-	Email             string `db:"email"`
-	PointsCollected   int64  `db:"points_collected"`
-	CashbackCollected int64  `db:"cashback_collected"`
+	ID                string  `db:"id"`
+	UserDNI           string  `db:"user_dni"`
+	Name              string  `db:"name"`
+	LastName          string  `db:"last_name"`
+	Phone             string  `db:"phone"`
+	Email             string  `db:"email"`
+	PointsCollected   float64 `db:"points_collected"`
+	CashbackCollected float64 `db:"cashback_collected"`
 }
